refactor: keep Rust files under 300 lines
The mise pin tests and a few older modules crossed the review limit. Split them so each file stays readable.
