+++ src/config/tests/mod.rs
@@ -1,3 +1,4 @@
 mod collect;
 mod hydrate;
 mod plan;
+mod plan_network;
