+++ src/inject/set.rs
@@ -11,10 +11,10 @@ use crate::shell::Shell;
 use crate::ticket;
 
 use super::acquire::acquire_secrets_and_network;
+use super::pins::{apply_pins, select_tool_env};
+use super::work::{SecretHydrate, pre_event_from_work, resolve_provider_work, ticket_ready};
 use super::{
-    SecretHydrate, apply_pins, emit_seen_if_walk_log, merge_env_with_conflicts,
-    pre_event_from_work, public_hydrate, resolve_provider_work, select_tool_env,
-    show_loader_warnings, ticket_ready,
+    emit_seen_if_walk_log, merge_env_with_conflicts, public_hydrate, show_loader_warnings,
 };
 
 pub async fn handle_set(
