+++ src/network/acquire.rs
@@ -5,15 +5,16 @@ use std::process::Command;
 use std::time::{Duration, Instant};
 
 use crate::config::NetworkBinding;
+use crate::network::labels::{
+    connection_label, env_entry_for, provider_label, send_failed, send_progress,
+};
 use crate::network::parse::{parse_binding, reconcile_local_port};
 use crate::network::process::{
     ChildOutputFiles, RunningForward, configure_child_process, stop_network_pids_list,
     wait_child_ready,
 };
-use crate::network::progress::{ProviderProgressEvent, ProviderProgressKind, format_timing};
-use crate::network::types::{
-    AcquiredNetwork, DetachedNetworkSession, LocalTarget, ParsedBinding, ProviderSpec,
-};
+use crate::network::progress::ProviderProgressKind;
+use crate::network::types::{AcquiredNetwork, DetachedNetworkSession, ParsedBinding};
 use crate::provider_progress::ProviderProgressSink;
 use lade_sdk::network::{build_command, ensure_provider_preflight};
 
@@ -266,75 +267,4 @@ fn acquire_detached_binding(
         ProviderProgressKind::Connected,
     );
     Ok((env_entry, pid))
-}
-
-fn env_entry_for(target: &LocalTarget, local_port: u16) -> Option<(String, String)> {
-    match target {
-        LocalTarget::EnvVar(name) => Some((name.clone(), local_port.to_string())),
-        LocalTarget::FixedPort(_) => None,
-    }
-}
-
-fn connection_label(spec: &ProviderSpec, local_host: &str, local_port: u16) -> String {
-    let local = if local_host == "127.0.0.1" || local_host == "localhost" {
-        local_port.to_string()
-    } else {
-        format!("{local_host}:{local_port}")
-    };
-    match spec {
-        ProviderSpec::Kubectl {
-            name, remote_port, ..
-        } => format!("{name}:{remote_port} on {local}"),
-        ProviderSpec::Kubefwd {
-            name, service_port, ..
-        } => format!("{name}:{service_port} on {local}"),
-        ProviderSpec::TshKubeCluster {
-            name, remote_port, ..
-        } => format!("{name}:{remote_port} on {local}"),
-        ProviderSpec::TshApp {
-            app_name,
-            target_port,
-            ..
-        } => match target_port {
-            Some(target_port) => format!("{app_name}:{target_port} on {local}"),
-            None => format!("{app_name} on {local}"),
-        },
-        ProviderSpec::Ssh {
-            remote_host,
-            remote_port,
-            ..
-        } => format!("{remote_host}:{remote_port} on {local}"),
-    }
-}
-
-fn provider_label(spec: &ProviderSpec) -> &'static str {
-    match spec {
-        ProviderSpec::Kubectl { .. } => "kubectl forward",
-        ProviderSpec::Kubefwd { .. } => "kubefwd forward",
-        ProviderSpec::TshKubeCluster { .. } => "tsh kube_cluster forward",
-        ProviderSpec::TshApp { .. } => "tsh app proxy",
-        ProviderSpec::Ssh { .. } => "ssh forward",
-    }
-}
-
-fn send_failed(progress: &ProviderProgressSink, id: String, display: String, started: Instant) {
-    send_progress(
-        progress,
-        &id,
-        format_timing(&display, started),
-        ProviderProgressKind::Failed,
-    );
-}
-
-fn send_progress(
-    progress: &ProviderProgressSink,
-    id: &str,
-    display: String,
-    kind: ProviderProgressKind,
-) {
-    progress.send(ProviderProgressEvent {
-        id: id.to_string(),
-        display,
-        kind,
-    });
 }
